requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
flake8==7.0.0
frozenlist==1.4.1
idna==3.7
isort==5.13.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
python-dotenv==1.0.1
typing_extensions==4.11.0
yarl==1.9.4
